加入token前

+++ shop/src/main/java/com/example/shop/repository/CarouselViewRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import com.example.shop.entity.CarouselView;
 import com.example.shop.entity.Indexconfig;
 @RepositoryRestResource
-public interface CarouselViewRepository extends JpaRepository<CarouselView, Integer> {
+public interface CarouselViewRepository extends JpaRepository<CarouselView, Long> {
 
 	List<CarouselView> findByIndexconfig(Indexconfig indexconfig);
 	List<CarouselView> findByIndexconfig(Integer indexconfig_id);
